requirements.txt
gspread
gspread-dataframe
oauth2client
pandas
requests
bs4
openpyxl
xlrd
browser-cookie3
pycookiecheat
selenium
webdrivermanager
#pysqlite
fuzzywuzzy
seaborn
nltk
sklearn
sumy
pdfminer.six
wordcloud
python-linkedin-v2
